perf: Putting the values as int and using .AsNoTracking (EF Core)

+++ RinhaBackend.Api/DTO/InsertTransaction/InputInsertTransactionDTO.cs
@@ -6,7 +6,7 @@ namespace RinhaBackend.Api.DTO.InsertTransaction
     public class InputInsertTransactionDTO()
     {
         [JsonPropertyName("Valor")]
-        public decimal Value { get; set; }
+        public int Value { get; set; }
 
         // TODO: Type must be "c" or "d"
         [JsonPropertyName("Tipo")]
